[feat]: 반응형 화면 구현 및 검색 버튼 추가
