+++ src/App.css
@@ -1,5 +1,9 @@
 .App {
   text-align: center;
-
+  width: 100%;
+  height: auto;
 }
 
+.search-button {
+  cursor: pointer;
+}
